add fill document from queue
