+++ apps/io-func-sign-user/src/app/use-cases/create-filled-document.ts
@@ -2,42 +2,35 @@ import * as E from "fp-ts/lib/Either";
 import * as TE from "fp-ts/lib/TaskEither";
 
 import { pipe } from "fp-ts/lib/function";
-import { GetFiscalCodeBySignerId, Signer } from "@internal/io-sign/signer";
+import { Signer } from "@internal/io-sign/signer";
 
 import { EntityNotFoundError } from "@internal/io-sign/error";
-import { Field, populatePdf } from "@internal/pdf-handler/pdf";
+
 import { validate } from "@internal/io-sign/validation";
 
 import { EmailString, NonEmptyString } from "@pagopa/ts-commons/lib/strings";
+import * as t from "io-ts";
 
 import { FilledDocumentUrl } from "../../filled-document";
-import { GetBlobUrl, UploadBlob } from "../../infra/azure/storage/blob";
+import { GetBlobUrl } from "../../infra/azure/storage/blob";
 import { EnqueueMessage } from "../../infra/azure/storage/queue";
 
-// these types define the fields inside the PDF file to be enhanced
-// TODO: These are not yet the real parameters. Pending communication from the QTSP [SFEQS-1164]
-type NameField = Field & { fieldName: "name" };
-type FamilyNameField = Field & { fieldName: "surname" };
-type EmailField = Field & { fieldName: "email" };
-type FiscalCodeField = Field & { fieldName: "CF" };
-type Fields = [NameField, FamilyNameField, EmailField, FiscalCodeField];
+export const CreateFilledDocumentPayload = t.type({
+  signer: Signer,
+  documentUrl: NonEmptyString,
+  email: EmailString,
+  familyName: NonEmptyString,
+  name: NonEmptyString,
+});
 
-export type PrepareFilledDocumentPayload = {
-  signer: Signer;
-  documentUrl: NonEmptyString;
-  email: EmailString;
-  familyName: NonEmptyString;
-  name: NonEmptyString;
-};
-
-export type CreateFilledDocumentPayload = PrepareFilledDocumentPayload & {
-  filledDocumentFileName: NonEmptyString;
-};
+export type CreateFilledDocumentPayload = t.TypeOf<
+  typeof CreateFilledDocumentPayload
+>;
 
 /* This function returns only the callback url of the filled document without creating it.
  * It also writes on a queue the information necessary to start the module creation function via the trigger queue
  */
-export const makePrepareFilledDocument =
+export const makeCreateFilledDocument =
   (getFilledDocumentUrl: GetBlobUrl, enqueueDocumentToFill: EnqueueMessage) =>
   ({
     signer,
@@ -45,7 +38,7 @@ export const makePrepareFilledDocument =
     email,
     familyName,
     name,
-  }: PrepareFilledDocumentPayload) => {
+  }: CreateFilledDocumentPayload) => {
     const filledDocumentFileName = `${signer.id}.pdf`;
 
     return pipe(
@@ -79,58 +72,3 @@ export const makePrepareFilledDocument =
       )
     );
   };
-
-// This function downloads the pdf form, compiles it and uploads it to blobStorage
-export const makeCreateFilledDocument =
-  (
-    getFiscalCodeBySignerId: GetFiscalCodeBySignerId,
-    uploadFilledDocument: UploadBlob,
-    fetchWithTimeout: typeof fetch
-  ) =>
-  ({
-    signer,
-    documentUrl,
-    email,
-    familyName,
-    name,
-    filledDocumentFileName,
-  }: CreateFilledDocumentPayload) =>
-    pipe(
-      signer.id,
-      getFiscalCodeBySignerId,
-      TE.chain(
-        TE.fromOption(
-          () =>
-            new EntityNotFoundError("Fiscal code not found for this signer!")
-        )
-      ),
-      TE.chain((fiscalCode) => {
-        const fields: Fields = [
-          {
-            fieldName: "name",
-            fieldValue: name,
-          },
-          {
-            fieldName: "surname",
-            fieldValue: familyName,
-          },
-          {
-            fieldName: "email",
-            fieldValue: email,
-          },
-          {
-            fieldName: "CF",
-            fieldValue: fiscalCode,
-          },
-        ];
-
-        return pipe(
-          TE.tryCatch(() => fetchWithTimeout(documentUrl), E.toError),
-          TE.chain((response) => TE.tryCatch(() => response.blob(), E.toError)),
-          TE.chain((blob) => TE.tryCatch(() => blob.arrayBuffer(), E.toError)),
-          TE.map((arrayBuffer) => Buffer.from(arrayBuffer)),
-          TE.chain((buffer) => pipe(fields, populatePdf(buffer))),
-          TE.chain(uploadFilledDocument(filledDocumentFileName))
-        );
-      })
-    );
